requirements.txt
# Core Framework
streamlit>=1.28.0
click>=8.1.0
aiohttp>=3.8.0
httpx>=0.25.0

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0

# Code Analysis
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0

# Storage & Data
chromadb>=0.4.0
pydantic>=2.0.0
PyYAML>=6.0

# Performance & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
hypothesis>=6.0.0
memory-profiler>=0.60.0

# Code Quality Tools
pylint>=2.17.0
flake8>=6.0.0
black>=23.0.0

# Utilities
watchdog>=3.0.0
rich>=13.0.0
plotly>=5.17.0

# GitHub Integration
PyGithub>=1.59.0

# Scientific Computing
scikit-learn>=1.3.0
numpy>=1.24.0